The finalisePath option has been deprecated in favour of the new finaliseDirectory option, which dictates the root directory of the finalise repository. This can be configured in an environment variable called "FinaliseDirectory".
Any git repositories under the root FinaliseDirectory will be purged, in theory as a one-off operation.

Zip Backups are no longer performed, using the finaliseDirectory option is the preferred alternative.

New function: TagVersion - adds a tag to the finalise repository with the specified version number

NGit library removed in preference of LibGit2Sharp which is more performant on large repositories.